pyproject.toml
[project]
name = "query-bier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beir>=2.2.0",
    "einops>=0.8.1",
    "fastapi>=0.116.1",
    "google-genai>=1.31.0",
    "huggingface-hub[cli]>=0.34.4",
    "loguru>=0.7.3",
    "openai>=1.102.0",
    "pandas>=2.3.2",
    "peft>=0.17.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "uvicorn>=0.35.0",
]
